+++ CHANGELOG.md
@@ -14,7 +14,8 @@ First ever release. You can make flows from JSON or YAML.
 - Issue #11: Created a file [DEVELOPMENT.md] to keep [README.md] end-user-friendly.
 - Issue #12: You can set the contents of the example buttons via the configs.
 - Issue #19: You can import and export figure configurations.
-- Issue #22: There is a button to import the sizes of all nodes of the flow.
+- Issue #22: There is a button to import the positions of all nodes of the flow.
+- Issue #30: There is a switch button to auto-sync all the changes you make in the editors.
 - Issue #34: You can use define for nodes and edges.
 - Issue #35: Nodes can have a dashed stroke (just like edges already could).
 - Issue #59: There is a button to export the raw ReactFlow config.